+++ src/brute.h
@@ -4,6 +4,6 @@
 #include <vector>
 #include <cstdint>
 
-void brute_force_solve(const std::vector<std::vector<int64_t>>& clause_list, const uint64_t variable_count);
+void brute_force_solve(const std::vector<std::vector<int32_t>>& clause_list, const uint32_t variable_count);
 
 #endif
