+++ src/file.cpp
@@ -14,14 +14,14 @@
  * and
  * https://fairmut3x.wordpress.com/2011/07/29/cnf-conjunctive-normal-form-dimacs-format-explained/
  */
-std::pair<std::vector<std::vector<int64_t>>, int64_t> read_file(const char *path) {
+std::pair<std::vector<std::vector<int32_t>>, int32_t> read_file(const char *path) {
     std::ifstream file{path};
 
     bool in_comments = true;
-    int64_t variables = -1;
-    int64_t clauses = -1;
+    int32_t variables = -1;
+    int32_t clauses = -1;
 
-    std::vector<std::vector<int64_t>> clause_list;
+    std::vector<std::vector<int32_t>> clause_list;
 
     for (std::string line; std::getline(file, line);) {
         //Ignore comments
@@ -46,13 +46,13 @@ std::pair<std::vector<std::vector<int64_t>>, int64_t> read_file(const char *path
             std::cerr << "Problem line was not found in file\n";
             return {};
         }
-        int64_t token_num;
-        std::vector<int64_t> clause_tokens;
+        int32_t token_num;
+        std::vector<int32_t> clause_tokens;
 
         std::istringstream iss{std::move(line)};
 
-        clause_tokens.assign(std::istream_iterator<int64_t>(iss),
-                std::istream_iterator<int64_t>());
+        clause_tokens.assign(std::istream_iterator<int32_t>(iss),
+                std::istream_iterator<int32_t>());
         clause_tokens.erase(std::remove(clause_tokens.begin(), clause_tokens.end(), 0),
                 clause_tokens.end());
         clause_tokens.shrink_to_fit();
